# Go: improve performance of first-party package analysis (Cherry-pick of #13476)