--- a/src/python/pants/backend/go/util_rules/BUILD
+++ b/src/python/pants/backend/go/util_rules/BUILD
@@ -1,7 +1,7 @@
 # Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-python_sources(dependencies=[":analyzer_sources"])
+python_sources(dependencies=[":go_sources"])
 python_tests(name="tests", timeout=120)
 
-resource(name="analyzer_sources", source="generate_testmain.go")
+resources(name="go_sources", sources=["*.go"])