--- a/src/python/pants/backend/go/util_rules/first_party_pkg.py
+++ b/src/python/pants/backend/go/util_rules/first_party_pkg.py
@@ -6,23 +6,23 @@ from __future__ import annotations
 import json
 import logging
 import os
+import pkgutil
 from dataclasses import dataclass
+from typing import ClassVar
 
 from pants.backend.go.target_types import (
     GoFirstPartyPackageSourcesField,
     GoFirstPartyPackageSubpathField,
     GoImportPathField,
 )
+from pants.backend.go.util_rules.build_pkg import BuildGoPackageRequest, BuiltGoPackage
 from pants.backend.go.util_rules.go_mod import GoModInfo, GoModInfoRequest
-from pants.backend.go.util_rules.sdk import GoSdkProcess
-from pants.backend.go.util_rules.third_party_pkg import (
-    AllThirdPartyPackages,
-    AllThirdPartyPackagesRequest,
-)
+from pants.backend.go.util_rules.import_analysis import ImportConfig, ImportConfigRequest
+from pants.backend.go.util_rules.link import LinkedGoBinary, LinkGoBinaryRequest
 from pants.build_graph.address import Address
 from pants.engine.engine_aware import EngineAwareParameter
-from pants.engine.fs import Digest, MergeDigests
-from pants.engine.process import FallibleProcessResult
+from pants.engine.fs import CreateDigest, Digest, FileContent, MergeDigests
+from pants.engine.process import FallibleProcessResult, Process
 from pants.engine.rules import Get, MultiGet, collect_rules, rule
 from pants.engine.target import HydratedSources, HydrateSourcesRequest, WrappedTarget
 
@@ -71,9 +71,15 @@ class FirstPartyPkgInfoRequest(EngineAwareParameter):
         return self.address.spec
 
 
+@dataclass(frozen=True)
+class PackageAnalyzerSetup:
+    digest: Digest
+    PATH: ClassVar[str] = "./_analyze_package"
+
+
 @rule
 async def compute_first_party_package_info(
-    request: FirstPartyPkgInfoRequest,
+    request: FirstPartyPkgInfoRequest, analyzer: PackageAnalyzerSetup
 ) -> FallibleFirstPartyPkgInfo:
     go_mod_address = request.address.maybe_convert_to_target_generator()
     wrapped_target, go_mod_info = await MultiGet(
@@ -84,24 +90,23 @@ async def compute_first_party_package_info(
     import_path = target[GoImportPathField].value
     subpath = target[GoFirstPartyPackageSubpathField].value
 
-    pkg_sources, all_third_party_packages = await MultiGet(
-        Get(HydratedSources, HydrateSourcesRequest(target[GoFirstPartyPackageSourcesField])),
-        Get(AllThirdPartyPackages, AllThirdPartyPackagesRequest(go_mod_info.stripped_digest)),
+    pkg_sources = await Get(
+        HydratedSources, HydrateSourcesRequest(target[GoFirstPartyPackageSourcesField])
     )
     input_digest = await Get(
         Digest,
-        MergeDigests(
-            [pkg_sources.snapshot.digest, go_mod_info.digest, all_third_party_packages.digest]
-        ),
+        MergeDigests([pkg_sources.snapshot.digest, analyzer.digest]),
     )
-
+    path = request.address.spec_path if request.address.spec_path else "."
+    path = os.path.join(path, subpath) if subpath else path
+    if not path:
+        path = "."
     result = await Get(
         FallibleProcessResult,
-        GoSdkProcess(
+        Process(
+            (analyzer.PATH, path),
             input_digest=input_digest,
-            command=("list", "-json", f"./{subpath}"),
             description=f"Determine metadata for {request.address}",
-            working_dir=request.address.spec_path,  # i.e. the `go.mod`'s directory.
         ),
     )
     if result.exit_code != 0:
@@ -113,6 +118,19 @@ async def compute_first_party_package_info(
         )
 
     metadata = json.loads(result.stdout)
+    if "Error" in metadata or "InvalidGoFiles" in metadata:
+        error = metadata.get("Error", "")
+        if error:
+            error += "\n"
+        if "InvalidGoFiles" in metadata:
+            error += "\n".join(
+                f"{filename}: {error}"
+                for filename, error in metadata.get("InvalidGoFiles", {}).items()
+            )
+            error += "\n"
+        return FallibleFirstPartyPkgInfo(
+            info=None, import_path=import_path, exit_code=1, stderr=error
+        )
 
     if "CgoFiles" in metadata:
         raise NotImplementedError(
@@ -135,6 +153,50 @@ async def compute_first_party_package_info(
         s_files=tuple(metadata.get("SFiles", [])),
     )
     return FallibleFirstPartyPkgInfo(info, import_path)
+
+
+@rule
+async def setup_analyzer() -> PackageAnalyzerSetup:
+    source_entry_content = pkgutil.get_data("pants.backend.go.util_rules", "analyze_package.go")
+    if not source_entry_content:
+        raise AssertionError("Unable to find resource for `analyze_package.go`.")
+
+    source_entry = FileContent("analyze_package.go", source_entry_content)
+
+    source_digest, import_config = await MultiGet(
+        Get(Digest, CreateDigest([source_entry])),
+        Get(ImportConfig, ImportConfigRequest, ImportConfigRequest.stdlib_only()),
+    )
+
+    built_analyzer_pkg = await Get(
+        BuiltGoPackage,
+        BuildGoPackageRequest(
+            import_path="main",
+            subpath="",
+            digest=source_digest,
+            go_file_names=(source_entry.path,),
+            s_file_names=(),
+            direct_dependencies=(),
+            minimum_go_version=None,
+        ),
+    )
+    main_pkg_a_file_path = built_analyzer_pkg.import_paths_to_pkg_a_files["main"]
+    input_digest = await Get(
+        Digest, MergeDigests([built_analyzer_pkg.digest, import_config.digest])
+    )
+
+    analyzer = await Get(
+        LinkedGoBinary,
+        LinkGoBinaryRequest(
+            input_digest=input_digest,
+            archives=(main_pkg_a_file_path,),
+            import_config_path=import_config.CONFIG_PATH,
+            output_filename=PackageAnalyzerSetup.PATH,
+            description="Link Go package analyzer",
+        ),
+    )
+
+    return PackageAnalyzerSetup(analyzer.digest)
 
 
 def rules():