--- a/src/python/pants/backend/go/util_rules/first_party_pkg.py
+++ b/src/python/pants/backend/go/util_rules/first_party_pkg.py
@@ -25,6 +25,7 @@ from pants.engine.fs import CreateDigest, Digest, FileContent, MergeDigests
 from pants.engine.process import FallibleProcessResult, Process
 from pants.engine.rules import Get, MultiGet, collect_rules, rule
 from pants.engine.target import HydratedSources, HydrateSourcesRequest, WrappedTarget
+from pants.util.logging import LogLevel
 
 logger = logging.getLogger(__name__)
 
@@ -107,6 +108,7 @@ async def compute_first_party_package_info(
             (analyzer.PATH, path),
             input_digest=input_digest,
             description=f"Determine metadata for {request.address}",
+            level=LogLevel.DEBUG,
         ),
     )
     if result.exit_code != 0: