--- a/src/python/pants/backend/go/util_rules/first_party_pkg.py
+++ b/src/python/pants/backend/go/util_rules/first_party_pkg.py
@@ -179,7 +179,6 @@ async def setup_analyzer() -> PackageAnalyzerSetup:
             go_file_names=(source_entry.path,),
             s_file_names=(),
             direct_dependencies=(),
-            minimum_go_version=None,
         ),
     )
     main_pkg_a_file_path = built_analyzer_pkg.import_paths_to_pkg_a_files["main"]